DNM: retrigger wheel builds for 0.6.0 on non-x86 arch
In the recent 0.6.0 the wheel builds for non-x86 architectures failed
because the install version importlib-metadata in the travis CI nodes
prevented pip from installing a compatible version with the version
required for keyring (which is required for twine to work). So after
the wheels were built the twine upload command to push them to pypi
failed. This commit is an attempt to workaround this and force the CI
jobs to run from 0.6.0 while manually upgrading all the effected
packages to their latest versions. This will be discarded before merging
since it removes the conditions for running and publishing wheel jobs.